Moved refresh display function to domUpdates.js
